+++ gwemopt/utils/__init__.py
@@ -11,16 +11,4 @@ from gwemopt.utils.pixels import (
     getRectanglePixels,
     getSquarePixels,
 )
-from gwemopt.utils.tile_utils import (
-    append_tile_epochs,
-    balance_tiles,
-    check_overlapping_tiles,
-    eject_tiles,
-    erase_unbalanced_tiles,
-    optimize_max_tiles,
-    order_by_observability,
-    slice_galaxy_tiles,
-    slice_map_tiles,
-    slice_number_tiles,
-)
 from gwemopt.utils.treasuremap import get_treasuremap_pointings
